Move client projection tests to projections test project Projections manager should run on ExtHttpEndPoint
