--- src/EventStore.Core.Tests/ClientAPI/Projections/TestWithNode.cs
@@ -1,40 +1,0 @@
-﻿using System;
-using EventStore.ClientAPI;
-using EventStore.ClientAPI.Common.Log;
-using EventStore.ClientAPI.Projections;
-using EventStore.Core.Tests.ClientAPI.Helpers;
-using EventStore.Core.Tests.Helpers;
-using NUnit.Framework;
-
-namespace EventStore.Core.Tests.ClientAPI.Projections
-{
-    [Category("LongRunning")]
-    public class TestWithNode : SpecificationWithDirectoryPerTestFixture
-    {
-        protected MiniNode _node;
-        protected ProjectionsManager _manager;
-
-        [TestFixtureSetUp]
-        public override void TestFixtureSetUp()
-        {
-            base.TestFixtureSetUp();
-            _node = new MiniNode(PathName);
-            _node.Start();
-            _manager = new ProjectionsManager(new NoopLogger(), _node.ExtHttpEndPoint, TimeSpan.FromSeconds(5));
-        }
-
-        [TestFixtureTearDown]
-        public override void TestFixtureTearDown()
-        {
-            _node.Shutdown();
-            base.TestFixtureTearDown();
-        }
-
-
-        protected virtual IEventStoreConnection BuildConnection(MiniNode node)
-        {
-            return TestConnection.Create(node.TcpEndPoint);
-        }
-
-    }
-}
