+++ src/EventStore.Core.Tests/EventStore.Core.Tests.csproj
@@ -91,8 +91,6 @@
     <Compile Include="ClientAPI\Embedded\deleting_persistent_subscription.cs" />
     <Compile Include="ClientAPI\Embedded\update_persistent_subscription.cs" />
     <Compile Include="ClientAPI\persistent_connect_integration_tests.cs" />
-    <Compile Include="ClientAPI\Projections\list_projections.cs" />
-    <Compile Include="ClientAPI\Projections\TestWithNode.cs" />
     <Compile Include="ClientAPI\read_allevents_backward_with_linkto_deleted_event.cs" />
     <Compile Include="ClientAPI\Embedded\appending_to_implicitly_created_stream.cs" />
     <Compile Include="ClientAPI\Embedded\appending_to_implicitly_created_stream_using_transaction.cs" />
