+++ src/EventStore.Projections.Core.Tests/ClientAPI/list_projections.cs
@@ -3,9 +3,9 @@ using System.Linq;
 using EventStore.ClientAPI.SystemData;
 using NUnit.Framework;
 
-namespace EventStore.Core.Tests.ClientAPI.Projections
+namespace EventStore.Projections.Core.Tests.ClientAPI
 {
-    public class list_projections : TestWithNode
+    public class list_projections : specification_with_standard_projections_runnning
     {
         const string TestProjection = "fromAll().when({$init: function (state, ev) {return {};},ConversationStarted: function (state, ev) {state.lastBatchSent = ev;return state;}});";
