+++ src/EventStore.Projections.Core.Tests/EventStore.Projections.Core.Tests.csproj
@@ -61,6 +61,7 @@
     <Compile Include="ClientAPI\Cluster\specification_with_standard_projections_runnning.cs" />
     <Compile Include="ClientAPI\event_by_type_index.cs" />
     <Compile Include="ClientAPI\event_by_type_index\when_reverting_after_index_catches_up.cs" />
+    <Compile Include="ClientAPI\list_projections.cs" />
     <Compile Include="ClientAPI\specification_with_standard_projections_runnning.cs" />
     <Compile Include="ClientAPI\when_handling_created\with_from_all_any_foreach_projection\when_running_and_events_are_posted.cs" />
     <Compile Include="ClientAPI\when_handling_created\with_from_all_foreach_projection\when_running_and_events_are_indexed.cs" />
